+++ src/ts/wysiwyg/processCodeRender.ts
@@ -1,3 +1,4 @@
+import codeSVG from "../../assets/icons/code.svg";
 import {setSelectionFocus} from "../editor/setSelection";
 import {abcRender} from "../markdown/abcRender";
 import {chartRender} from "../markdown/chartRender";
@@ -5,7 +6,6 @@ import {codeRender} from "../markdown/codeRender";
 import {highlightRender} from "../markdown/highlightRender";
 import {mathRenderByLute} from "../markdown/mathRenderByLute";
 import {mermaidRender} from "../markdown/mermaidRender";
-import codeSVG from "../../assets/icons/code.svg";
 
 // html, math, math-inline, code block, abc, chart, mermaid
 export const processCodeRender = (blockElement: HTMLElement, vditor: IVditor) => {
@@ -61,9 +61,9 @@ export const processCodeRender = (blockElement: HTMLElement, vditor: IVditor) =>
             codeRender(previewPanel, vditor.options.lang);
         }
     } else if (blockType.indexOf("html") > -1) {
-        if (blockType === 'html-inline') {
+        if (blockType === "html-inline") {
             previewPanel.innerHTML = codeSVG;
-            return
+            return;
         }
         previewPanel.innerHTML = innerHTML.replace(/&amp;/g, "&")
             .replace(/&lt;/g, "<").replace(/&gt;/g, ">");